Feature/019 architectural updates rx (#14)
* 019 architectural improvements
* 019 retrofit builder made more general
* 019 minor documentation cleanup
* Create LICENSE.md
