+++ app/src/main/java/com/guidovezzoni/bingeworthyshow2/data/api/MdbApiInstance.kt
@@ -1,0 +1,12 @@
+package com.guidovezzoni.bingeworthyshow2.data.api
+
+import com.guidovezzoni.bingeworthyshow2.utils.networking.RetrofitBuilder
+
+object MdbApiInstance {
+    private const val BASE_URL = "https://api.themoviedb.org"
+
+    private fun getMdbApi(): MdbApi =
+        RetrofitBuilder.getRetrofit(BASE_URL).create(MdbApi::class.java)
+
+    val MDB_API: MdbApi = getMdbApi()
+}
