+++ app/src/main/java/com/guidovezzoni/bingeworthyshow2/data/di/DiDataProvider.kt
@@ -1,17 +1,14 @@
 package com.guidovezzoni.bingeworthyshow2.data.di
 
 import com.guidovezzoni.bingeworthyshow2.BuildConfig.THE_MOVIE_DB_KEY
-import com.guidovezzoni.bingeworthyshow2.data.api.MdbApi
+import com.guidovezzoni.bingeworthyshow2.data.api.MdbApiInstance.MDB_API
 import com.guidovezzoni.bingeworthyshow2.data.datasource.MdbRestDatasource
-import com.guidovezzoni.bingeworthyshow2.data.repository.MdbRepository
 import com.guidovezzoni.bingeworthyshow2.data.repository.MdbRepositoryImpl
-import com.guidovezzoni.bingeworthyshow2.utils.networking.RetrofitBuilder
+import com.guidovezzoni.bingeworthyshow2.domain.repository.MdbRepository
 
-object DiProvider {
-    val mdbApi: MdbApi by lazy { RetrofitBuilder.MDB_API }
-
+object DiDataProvider {
     val mdbRepository: MdbRepository by lazy {
-        val datasource = MdbRestDatasource(mdbApi, THE_MOVIE_DB_KEY)
+        val datasource = MdbRestDatasource(MDB_API, THE_MOVIE_DB_KEY)
         MdbRepositoryImpl(datasource)
     }
 }
