+++ app/src/main/java/com/guidovezzoni/bingeworthyshow2/domain/usecase/GetConfigurationUseCase.kt
@@ -1,8 +1,8 @@
 package com.guidovezzoni.bingeworthyshow2.domain.usecase
 
-import com.guidovezzoni.bingeworthyshow2.data.repository.MdbRepository
+import com.guidovezzoni.bingeworthyshow2.data.dto.toConfigurationDomainModel
 import com.guidovezzoni.bingeworthyshow2.domain.model.ConfigurationDomainModel
-import com.guidovezzoni.bingeworthyshow2.domain.model.toConfigurationDomainModel
+import com.guidovezzoni.bingeworthyshow2.domain.repository.MdbRepository
 import io.reactivex.rxjava3.core.Observable
 
 class GetConfigurationUseCase(private val repository: MdbRepository) {
