+++ app/src/main/java/com/guidovezzoni/bingeworthyshow2/domain/usecase/GetTopRatedShowsUseCase.kt
@@ -1,9 +1,9 @@
 package com.guidovezzoni.bingeworthyshow2.domain.usecase
 
-import com.guidovezzoni.bingeworthyshow2.data.repository.MdbRepository
+import com.guidovezzoni.bingeworthyshow2.data.dto.toPaginatedTvShowList
 import com.guidovezzoni.bingeworthyshow2.domain.model.PaginatedListDomainModel
 import com.guidovezzoni.bingeworthyshow2.domain.model.TvShowDomainModel
-import com.guidovezzoni.bingeworthyshow2.domain.model.toPaginatedTvShowList
+import com.guidovezzoni.bingeworthyshow2.domain.repository.MdbRepository
 import io.reactivex.rxjava3.core.Observable
 
 class GetTopRatedShowsUseCase(private val repository: MdbRepository) {
