+++ app/src/main/java/com/guidovezzoni/bingeworthyshow2/presentation/tvshowlist/viewmodel/ViewModelFactory.kt
@@ -2,8 +2,8 @@ package com.guidovezzoni.bingeworthyshow2.presentation.tvshowlist.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.guidovezzoni.bingeworthyshow2.data.api.MdbApi
-import com.guidovezzoni.bingeworthyshow2.domain.di.DiProvider
+import com.guidovezzoni.bingeworthyshow2.data.di.DiDataProvider
+import com.guidovezzoni.bingeworthyshow2.domain.di.DiDomainProvider
 
 class ViewModelFactory() :
     ViewModelProvider.Factory {
@@ -11,8 +11,8 @@ class ViewModelFactory() :
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         if (modelClass.isAssignableFrom(TvShowListViewModel::class.java)) {
             return TvShowListViewModel(
-                DiProvider.provideGetConfigurationUseCase(),
-                DiProvider.provideGetTopRatedShowsUseCase(),
+                DiDomainProvider.provideGetConfigurationUseCase(DiDataProvider.mdbRepository),
+                DiDomainProvider.provideGetTopRatedShowsUseCase(DiDataProvider.mdbRepository),
             ) as T
         }
         throw IllegalArgumentException("Unknown class name")
