+++ build.gradle
@@ -26,7 +26,7 @@ ext {
     ANDROID_COMPILE_SDK_VERSION = 33
 
     // google
-    GOOGLE_MATERIAL_VERSION = '1.6.1'
+    GOOGLE_MATERIAL_VERSION = '1.7.0'
 
     // androidX
     ANDROIDX_CORE_VERSION = "1.9.0"
@@ -44,6 +44,8 @@ ext {
     JUNIT_4_VERSION = "4.13.2"
     JUNIT_5_VERSION = "5.7.1"
     MOCKK_VERSION = "1.12.1"
-    ANDROIDX_TEST_VERSION = "1.1.3"
-    ANDROIDX_ESPRESSO_VERSION = "3.4.0"
+    ANDROIDX_TEST_VERSION = "1.1.4"
+    ANDROIDX_ESPRESSO_VERSION = "3.5.0"
+    KOTLIN_COROUTINE_TEST_VERSION = "1.6.0"
+    ANDROIDX_ARCH_CORE_TESTING_VERSION= "2.1.0"
 }
